Refactored by extracting r8 functionality as a trait
