+++ libs/androidlib/src/mill/androidlib/AndroidBuildTypeSettings.scala
@@ -14,7 +14,6 @@ case class AndroidBuildTypeSettings(
     isMinifyEnabled: Boolean = false,
     isShrinkEnabled: Boolean = false,
     enableDesugaring: Boolean = true,
-    useR8: Boolean = false,
     proguardFiles: ProguardFiles = ProguardFiles()
 ) {
   def withProguardLocalFiles(localFiles: Seq[os.Path]): AndroidBuildTypeSettings =
